+++ client/src/components/Signup/SignupSection/SignupSection.styles.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const S_PasswordGuide = styled.p`
+  font-size: 12px;
+  margin-bottom: 20px;
+  color: var(--color-primary-black);
+`;
